scrape/requirements.txt
boto3
requests
selenium
beautifulsoup4
dataclasses
dataclasses-json
python-dotenv